認証エラーでページが停止した場合にログアウトボタンを表示する機能を実装
